LPS-127341 Add A11y component and scheduler engine
TL;DR is an implementation for executing axe-core in real-time at runtime, configurable to observe changes in the UI and add overlays to elements with accessibility errors.

The Scheduler uses heuristic algorithms to be more responsive and to manage callback scheduling that is added by DOM mutations that need to be analyzed.